Workaround for Wayland multi-monitor windows size & pos issue
Requires GTK 3.22 in the current state, though it
should be possible to backport the changes to 3.0.

Wayland does not have global pointer coordinate system which made
previous algorithm for determining the screen with the cursor useless.
Furthermore window.move does not work for the same reason (at least for me).

As Wayland always places the window to be shown on the "active" monitor
(the one with the cursor) and all my attempts to change that failed,
I decided to at least tackle the window size issue.
Again we cannot determine monitor from neither cursor nor window position
(as the first would be relative to window and second to the active monitor
- or window, I am not entirely sure, but I have tried that), but we can
use our window to check on which monitor it is (probably) placed.

All we need to do is to maximize it and unmaximize it quickly
and measure the size in the meanwhile. That's all that this fix
strives to achieve. The fix does not detect Wayland nor care about
the GTK 3.0 compatibility but it might be a goof
starting point for a better solution.

The idea of using maximized window was inspired by:
https://stackoverflow.com/a/11175074/6646912 (thanks!)

The workaround might introduce a small lag when opening
the guake window but it is fast enough for me.